update install path, now in libexec
the package is kept Multi-Arch:foreign even though technically it is now
co-installable. since kconfig_compiler_kf5 is a tool used during build it
really needs to be able to use the native architecture as it is essentially
part of an x-compile toolchain. the only multiarch tag that affords the
dependency resolver the ability to prefer native is foreign so we really
should go with foreign for now to enable proper x-compile. in particular
otherwise one would have to manually install bin-dev:amd64 which is a bit
meh